refactor: move pipeline/routing.rs to astra-turn-core (828 LOC)
Move the unified routing decision engine (RoutingDecision, RoutingEngine) to
astra-turn-core where its deps (ConversationState, routing_metrics) already live.

Changes:
- Copy routing.rs to astra-turn-core::routing_engine (828 LOC)
- Register module in lib.rs
- Replace original with re-export stub
- Fix imports to use crate:: paths
